initial tuner support
